+++ src/main/java/cn/thinkmoon/blog/modules/controller/PostController.java
@@ -2,10 +2,10 @@ package cn.thinkmoon.blog.modules.controller;
 
 import cn.thinkmoon.blog.core.enums.PermissionTag;
 import cn.thinkmoon.blog.core.annotation.Permission;
-import cn.thinkmoon.blog.modules.pojo.UserPO;
-import cn.thinkmoon.blog.modules.pojo.PostPO;
+import cn.thinkmoon.blog.modules.pojo.po.UserPO;
+import cn.thinkmoon.blog.modules.pojo.po.PostPO;
 import cn.thinkmoon.blog.modules.service.PostService;
-import cn.thinkmoon.blog.modules.vo.PostVo;
+import cn.thinkmoon.blog.modules.pojo.vo.PostVO;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import lombok.extern.slf4j.Slf4j;
@@ -38,7 +38,7 @@ public class PostController {
     @Permission(permissionTag = PermissionTag.ADMIN)
     @RequestMapping(value = "/add")
     public int addPost(
-            @RequestBody PostVo postVo,
+            @RequestBody PostVO postVo,
             @RequestAttribute("user_info") UserPO user) {
         return postService.addPost((Integer) user.getId(), postVo.getTitle(), postVo.getText(), postVo.getCategory_id(), postVo.getCustomOptions());
     }
@@ -46,7 +46,7 @@ public class PostController {
     @Permission(permissionTag = PermissionTag.ADMIN)
     @RequestMapping(value = "/update")
     public int updatePost(
-            @RequestBody PostVo postVo,
+            @RequestBody PostVO postVo,
             @RequestAttribute("user_info") UserPO user) {
         return postService.updatePost(postVo.getCid(),user.getId(), postVo.getTitle(), postVo.getText(), postVo.getCategory_id(),postVo.getCustomOptions());
     }
