+++ src/main/java/cn/thinkmoon/blog/modules/controller/TagController.java
@@ -1,6 +1,6 @@
 package cn.thinkmoon.blog.modules.controller;
 
-import cn.thinkmoon.blog.modules.pojo.TagPO;
+import cn.thinkmoon.blog.modules.pojo.po.TagPO;
 import cn.thinkmoon.blog.modules.service.TagService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
