+++ src/main/java/cn/thinkmoon/blog/modules/pojo/dao/CategoryDAO.java
@@ -1,6 +1,6 @@
-package cn.thinkmoon.blog.modules.dao;
+package cn.thinkmoon.blog.modules.pojo.dao;
 
-import cn.thinkmoon.blog.modules.pojo.CategoryPO;
+import cn.thinkmoon.blog.modules.pojo.po.CategoryPO;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 
 import java.util.List;
