+++ src/main/java/cn/thinkmoon/blog/modules/pojo/dao/UserDAO.java
@@ -1,6 +1,6 @@
-package cn.thinkmoon.blog.modules.dao;
+package cn.thinkmoon.blog.modules.pojo.dao;
 
-import cn.thinkmoon.blog.modules.pojo.UserPO;
+import cn.thinkmoon.blog.modules.pojo.po.UserPO;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 
 public interface UserDAO extends BaseMapper<UserPO> {
