+++ src/main/java/cn/thinkmoon/blog/modules/service/SiteMapService.java
@@ -1,8 +1,6 @@
 package cn.thinkmoon.blog.modules.service;
 
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
+import cn.thinkmoon.blog.modules.pojo.po.SiteMapPO;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
@@ -12,26 +10,19 @@ public class SiteMapService {
             "<urlset xmlns=\"http://www.sitemaps.org/schemas/sitemap/0.9\">";
 
     static String END_DOC = "</urlset>";
-    static String CHANGEFREQ_DAILY = "daily";
-    public static String CHANGEFREQ_ALWAYS = "always";
+    public static String BASE_URL = "https://www.thinkmoon.c/";
 
 
-    @GetMapping(value = "/sitemap.xml", produces = {"application/xml;charset=UTF-8"})
-    @ResponseBody
-    public void getSiteMap(HttpServletResponse response) throws IOException {
+    public StringBuffer getSiteMap(HttpServletResponse response) throws IOException {
         StringBuffer sb = new StringBuffer();
         sb.append(BEGIN_DOC);//拼接开始部分
-        sb.append(new SiteMap("https://www.lovelin.space/"));//拼接网站首页地址
-        //下面是根据实际情况写，目的是生成整站的Url
-//        List<String> param = new ArrayList<>();
-        List<Blog> list=  blogMapper.findAllBlog();
+        sb.append(new SiteMapPO(BASE_URL));//拼接网站首页地址
+
+        List<Blog> list = blogMapper.findAllBlog();
         for (Blog blog : list) {
-            sb.append(new SiteMap("https://www.lovelin.space/blog/detail/" +blog.getBid() +".html" ,blog.getCreateTime(), CHANGEFREQ_DAILY, "0.8"));
+            sb.append(new SiteMap("https://www.lovelin.space/blog/detail/" + blog.getBid() + ".html", blog.getCreateTime(), CHANGEFREQ_DAILY, "0.8"));
         }
         sb.append(END_DOC);//拼接结尾
-        response.setContentType(MediaType.APPLICATION_XML_VALUE);
-        Writer writer = response.getWriter();
-
-        writer.append(sb.toString());
+        return sb;
     }
 }
