+++ src/main/java/cn/thinkmoon/blog/modules/service/TagService.java
@@ -1,7 +1,7 @@
 package cn.thinkmoon.blog.modules.service;
 
-import cn.thinkmoon.blog.modules.pojo.TagPO;
-import cn.thinkmoon.blog.modules.dao.TagDAO;
+import cn.thinkmoon.blog.modules.pojo.po.TagPO;
+import cn.thinkmoon.blog.modules.pojo.dao.TagDAO;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
