+++ src/main/java/cn/thinkmoon/blog/modules/service/UserService.java
@@ -1,7 +1,7 @@
 package cn.thinkmoon.blog.modules.service;
 
-import cn.thinkmoon.blog.modules.pojo.UserPO;
-import cn.thinkmoon.blog.modules.dao.UserDAO;
+import cn.thinkmoon.blog.modules.pojo.po.UserPO;
+import cn.thinkmoon.blog.modules.pojo.dao.UserDAO;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
