+++ src/main/resources/mapper/FieldMapper.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="cn.thinkmoon.blog.modules.dao.FieldDAO">
+<mapper namespace="cn.thinkmoon.blog.modules.pojo.dao.FieldDAO">
     <insert id="addField">
         INSERT INTO weblog_fields (cid,name,value)
         VALUES
